+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.2]
+
+- Add support for resizing SVG `<image>` tags alongside existing HTML `<img>` tags
+- Preserve `href` and `xlink:href` image sources when updating SVG image elements
+
 ## [1.0.1]
 
 - Update extension icon background to transparent